nodegl: add ngfx backend

+++ configure.py
@@ -35,8 +35,9 @@ import tarfile
 import urllib.request
 import venv
 import zipfile
-from multiprocessing import Pool
+from multiprocessing import Pool, cpu_count
 from subprocess import run
+import ssl
 
 
 _ROOTDIR = op.abspath(op.dirname(__file__))
@@ -76,6 +77,10 @@ _EXTERNAL_DEPS = dict(
         url='https://github.com/KhronosGroup/MoltenVK/archive/v@VERSION@.tar.gz',
         sha256='c20758bc19a46060aaf6e0949b47d29824b70b9ec0e22fb73a3feeef4c73a0ef',
     ),
+    ngfx=dict(
+        url='https://github.com/gopro/ngfx.git',
+        branch='develop',
+    ),
 )
 
 
@@ -88,6 +93,8 @@ def _get_external_deps(args):
             deps.append(_RENDERDOC_ID)
     if _SYSTEM == 'Darwin':
         deps += ['shaderc', 'moltenvk']
+    if args.enable_ngfx_backend:
+        deps += ['shaderc', 'ngfx']
     return {dep: _EXTERNAL_DEPS[dep] for dep in deps}
 
 
@@ -115,21 +122,39 @@ def _download_extract(dep_item):
 
     name, dep = dep_item
 
-    version = dep['version']
+    version = dep['version'] if 'version' in dep else ''
     url = dep['url'].replace('@VERSION@', version)
-    chksum = dep['sha256']
+    chksum = dep.get('sha256')
+    is_git_repo = url.endswith('.git')
+    if is_git_repo:
+        git_branch = dep['branch'] if 'branch' in dep else 'main'
     dst_file = dep.get('dst_file', op.basename(url)).replace('@VERSION@', version)
+    if dst_file.endswith('.git'):
+        dst_file = dst_file[:-4]
     dst_base = op.join(_ROOTDIR, 'external')
     dst_path = op.join(dst_base, dst_file)
     os.makedirs(dst_base, exist_ok=True)
 
     # Download
-    if not op.exists(dst_path) or not _file_chk(dst_path, dep['sha256']):
-        logging.info('downloading %s to %s', url, dst_file)
-        urllib.request.urlretrieve(url, dst_path)
+    if is_git_repo:
+        if not op.exists(dst_path):
+            logging.info('cloning %s', url)
+            run(['git', 'clone', url, '-b', git_branch, dst_path], check=True)
+    else:
+        if not op.exists(dst_path) or not _file_chk(dst_path, dep['sha256']):
+            logging.info('downloading %s to %s', url, dst_file)
+            # Temp workaround for pkgconf ssl verify error
+            ctx = ssl.create_default_context()
+            ctx.check_hostname = False
+            ctx.verify_mode = ssl.CERT_NONE
+            with urllib.request.urlopen(url, context=ctx) as u, open(dst_path, 'wb') as f:
+                f.write(u.read())
 
     # Extract
-    if tarfile.is_tarfile(dst_path):
+    # FIXME
+    if is_git_repo:
+        pass
+    elif tarfile.is_tarfile(dst_path):
         with tarfile.open(dst_path) as tar:
             dirs = {f.name for f in tar.getmembers() if f.isdir()}
             extract_dir = op.join(dst_base, _guess_base_dir(dirs))
@@ -148,17 +173,24 @@ def _download_extract(dep_item):
         assert False
 
     # Remove previous link if needed
+    # FIXME
     target = op.join(dst_base, name)
-    rel_extract_dir = op.basename(extract_dir)
-    if op.islink(target) and os.readlink(target) != rel_extract_dir:
-        logging.info('unlink %s target', target)
-        os.unlink(target)
-    elif op.exists(target) and not op.islink(target):
-        logging.info('remove previous %s copy', target)
-        shutil.rmtree(target)
+    if is_git_repo:
+        pass
+    else:
+        rel_extract_dir = op.basename(extract_dir)
+        if op.islink(target) and os.readlink(target) != rel_extract_dir:
+            logging.info('unlink %s target', target)
+            os.unlink(target)
+        elif op.exists(target) and not op.islink(target):
+            logging.info('remove previous %s copy', target)
+            shutil.rmtree(target)
 
     # Link (or copy)
-    if not op.exists(target):
+    # FIXME
+    if is_git_repo:
+        pass
+    elif not op.exists(target):
         logging.info('symlink %s -> %s', target, rel_extract_dir)
         try:
             os.symlink(rel_extract_dir, target)
@@ -184,19 +216,43 @@ def _block(name, prerequisites=None):
         return block_func
     return real_decorator
 
+def _get_cmake_setup_options(cfg, build_type = '$(CMAKE_BUILD_TYPE)', generator = '$(CMAKE_GENERATOR)', prefix = '$(PREFIX)', external_dir = '$(EXTERNAL_DIR)'):
+    opts = f'-DCMAKE_BUILD_TYPE={build_type} -G\"{generator}\" -DCMAKE_INSTALL_PREFIX=\"{prefix}\" -DEXTERNAL_DIR=\"{external_dir}\"'
+    if _SYSTEM == 'Windows':
+        opts += f' -DCMAKE_INSTALL_INCLUDEDIR=Include -DCMAKE_INSTALL_LIBDIR=Lib -DCMAKE_INSTALL_BINDIR=Scripts'
+        # Always use MultiThreadedDLL (/MD), not MultiThreadedDebugDLL (/MDd)
+        # Some external libraries are only available in Release mode, not in Debug mode
+        # MSVC toolchain doesn't allow mixing libraries built in /MD mode with libraries built in /MDd mode
+        opts += f' -DCMAKE_MSVC_RUNTIME_LIBRARY:STRING=MultiThreadedDLL'
+        # Set Windows SDK Version
+        opts += f' -DCMAKE_SYSTEM_VERSION=$(CMAKE_SYSTEM_VERSION)'
+        # Set VCPKG Directory
+        opts += f' -DVCPKG_DIR=$(VCPKG_DIR)'
+    return opts
+
+def _get_cmake_compile_options(cfg, build_type = '$(CMAKE_BUILD_TYPE)', num_threads = cpu_count() + 1):
+    opts = f'--config {build_type} -j{num_threads}'
+    if cfg.args.verbose:
+        opts += '-v'
+    return opts
+
+def _get_cmake_install_options(cfg, build_type = '$(CMAKE_BUILD_TYPE)'):
+    return f'--config {build_type}'
 
 def _meson_compile_install_cmd(component):
-    builddir = op.join('builddir', component)
+    builddir = op.join('$(BUILDDIR)', component)
     return ['$(MESON) ' + _cmd_join(action, '-C', builddir) for action in ('compile', 'install')]
 
 
 @_block('pkgconf-setup')
 def _pkgconf_setup(cfg):
-    return ['$(MESON_SETUP) ' + _cmd_join('-Dtests=false', cfg.externals['pkgconf'], op.join('builddir', 'pkgconf'))]
+    return ['$(MESON_SETUP) ' + _cmd_join('-Dtests=false', cfg.externals['pkgconf'], op.join('$(BUILDDIR)', 'pkgconf'))]
 
 
-@_block('pkgconf-install', [_pkgconf_setup])
+@_block('pkg-conf-install', [_pkgconf_setup])
 def _pkgconf_install(cfg):
+    if _SYSTEM != 'Windows':
+        return []
     ret = _meson_compile_install_cmd('pkgconf')
     pkgconf_exe = op.join(cfg.bin_path, 'pkgconf.exe')
     pkgconfig_exe = op.join(cfg.bin_path, 'pkg-config.exe')
@@ -205,7 +261,7 @@ def _pkgconf_install(cfg):
 
 @_block('sxplayer-setup')
 def _sxplayer_setup(cfg):
-    return ['$(MESON_SETUP) ' + _cmd_join(cfg.externals['sxplayer'], op.join('builddir', 'sxplayer'))]
+    return ['$(MESON_SETUP) ' + _cmd_join(cfg.externals['sxplayer'], op.join('$(BUILDDIR)', 'sxplayer'))]
 
 
 @_block('sxplayer-install', [_sxplayer_setup])
@@ -218,19 +274,64 @@ def _renderdoc_install(cfg):
     renderdoc_dll = op.join(cfg.externals[_RENDERDOC_ID], 'renderdoc.dll')
     return [f'copy {renderdoc_dll} {cfg.bin_path}']
 
+@_block('shaderc-install')
+def _shaderc_install(cfg):
+    shaderc_cmake_setup_options = _get_cmake_setup_options(cfg, build_type='Release', generator='Ninja')
+    shaderc_cmake_compile_options = _get_cmake_compile_options(cfg, build_type='Release')
+    shaderc_cmake_install_options = _get_cmake_install_options(cfg, build_type='Release')
+    if _SYSTEM == 'Darwin':
+        shaderc_lib_filename = 'libshaderc_shared.1.dylib'
+    cmd = []
+    if _SYSTEM in ['Darwin', 'Windows']:
+        cmd += [ 'cd external/shaderc && python ./utils/git-sync-deps' ]
+        cmd += [
+            f'$(CMAKE) -S external/shaderc -B $(BUILDDIR)/shaderc {shaderc_cmake_setup_options} -DSHADERC_SKIP_TESTS=ON',
+            f'$(CMAKE) --build $(BUILDDIR)/shaderc {shaderc_cmake_compile_options}',
+            f'$(CMAKE) --install $(BUILDDIR)/shaderc {shaderc_cmake_install_options}'
+        ]
+        if _SYSTEM == 'Darwin':
+            cmd += [ f'install_name_tool -id @rpath/{shaderc_lib_filename} $(PREFIX)/lib/{shaderc_lib_filename}' ]
+    return cmd
+
+@_block('ngfx-install', [_shaderc_install])
+def _ngfx_install(cfg):
+    ngfx_deps = ['nlohmann-json', 'stb']
+    if _SYSTEM == 'Windows':
+        ngfx_deps += ['glm', 'd3dx12', 'glfw']
+    ngfx_deps_str = ' '.join(s for s in ngfx_deps)
+    if _SYSTEM == 'Windows':
+        ngfx_install_deps_cmd = f'echo& set OS={_SYSTEM}& set PKGS={ngfx_deps_str}& python external/ngfx/build_scripts/install_deps.py&'
+    else:
+        ngfx_install_deps_cmd = f'export OS={_SYSTEM} && export PKGS="{ngfx_deps_str}" && python external/ngfx/build_scripts/install_deps.py'
+    cmd = [ ngfx_install_deps_cmd ]
+    ngfx_cmake_setup_options = _get_cmake_setup_options(cfg)
+    ngfx_cmake_compile_options = _get_cmake_compile_options(cfg)
+    ngfx_cmake_install_options = _get_cmake_install_options(cfg)
+    cmd += [
+        f'$(CMAKE) -S external/ngfx -B $(BUILDDIR)/ngfx {ngfx_cmake_setup_options} -D$(NGFX_GRAPHICS_BACKEND)=ON',
+        f'$(CMAKE) --build $(BUILDDIR)/ngfx --verbose {ngfx_cmake_compile_options}',
+        f'$(CMAKE) --install $(BUILDDIR)/ngfx {ngfx_cmake_install_options}'
+    ]
+    if cfg.ngfx_graphics_backend == 'NGFX_GRAPHICS_BACKEND_DIRECT3D12':
+        cmd += [ '$(PREFIX)\\Scripts\\ngfx_compile_shaders_dx12.exe d3dBlitOp' ]
+    return cmd
 
 @_block('nodegl-setup', [_sxplayer_install])
 def _nodegl_setup(cfg):
+    nodegl_setup_opts = ['--default-library', 'shared']
     nodegl_debug_opts = []
     if cfg.args.debug_opts:
         debug_opts = ','.join(cfg.args.debug_opts)
         nodegl_debug_opts += [f'-Ddebug_opts={debug_opts}']
+        if 'gpu_capture' in cfg.args.debug_opts:
+            renderdoc_dir = cfg.externals[_RENDERDOC_ID]
+            nodegl_debug_opts += [f'-Drenderdoc_dir={renderdoc_dir}']
+        nodegl_setup_opts += nodegl_debug_opts
 
-    if 'gpu_capture' in cfg.args.debug_opts and _SYSTEM != 'Darwin':
-        renderdoc_dir = cfg.externals[_RENDERDOC_ID]
-        nodegl_debug_opts += [f'-Drenderdoc_dir={renderdoc_dir}']
+    if cfg.args.enable_ngfx_backend:
+        nodegl_setup_opts += [f'-Dngfx_graphics_backend=$(NGFX_GRAPHICS_BACKEND)']
 
-    return ['$(MESON_SETUP) ' + _cmd_join(*nodegl_debug_opts, 'libnodegl', op.join('builddir', 'libnodegl'))]
+    return ['$(MESON_SETUP) ' + _cmd_join(*nodegl_setup_opts, 'libnodegl', op.join('$(BUILDDIR)', 'libnodegl'))]
 
 
 @_block('nodegl-install', [_nodegl_setup])
@@ -277,7 +378,7 @@ def _pynodegl_utils_install(cfg):
 
 @_block('ngl-tools-setup', [_nodegl_install])
 def _ngl_tools_setup(cfg):
-    return ['$(MESON_SETUP) ' + _cmd_join('ngl-tools', op.join('builddir', 'ngl-tools'))]
+    return ['$(MESON_SETUP) ' + _cmd_join('ngl-tools', op.join('$(BUILDDIR)', 'ngl-tools'))]
 
 
 @_block('ngl-tools-install', [_ngl_tools_setup])
@@ -286,7 +387,7 @@ def _ngl_tools_install(cfg):
 
 
 def _nodegl_run_target_cmd(cfg, target):
-    builddir = op.join('builddir', 'libnodegl')
+    builddir = op.join('$(BUILDDIR)', 'libnodegl')
     return ['$(MESON) ' + _cmd_join('compile', '-C', builddir, target)]
 
 
@@ -318,12 +419,12 @@ def _all(cfg):
 
 @_block('tests-setup', [_ngl_tools_install, _pynodegl_utils_install])
 def _tests_setup(cfg):
-    return ['$(MESON_SETUP_TESTS) ' + _cmd_join('tests', op.join('builddir', 'tests'))]
+    return ['$(MESON_SETUP_TESTS) ' + _cmd_join('tests', op.join('$(BUILDDIR)', 'tests'))]
 
 
 @_block('nodegl-tests', [_nodegl_install])
 def _nodegl_tests(cfg):
-    return [_cmd_join('meson', 'test', '-C', op.join('builddir', 'libnodegl'))]
+    return [_cmd_join('meson', 'test', '-C', op.join('$(BUILDDIR)', 'libnodegl'))]
 
 
 def _rm(f):
@@ -351,20 +452,20 @@ def _clean_py(cfg):
 @_block('clean', [_clean_py])
 def _clean(cfg):
     return [
-        _rd(op.join('builddir', 'libnodegl')),
-        _rd(op.join('builddir', 'moltenvk')),
-        _rd(op.join('builddir', 'ngl-tools')),
-        _rd(op.join('builddir', 'pkgconf')),  # XXX backport
-        _rd(op.join('builddir', 'shaderc')),
-        _rd(op.join('builddir', 'sxplayer')),
-        _rd(op.join('builddir', 'tests')),
+        _rd(op.join('$(BUILDDIR)', 'libnodegl')),
+        _rd(op.join('$(BUILDDIR)', 'moltenvk')),
+        _rd(op.join('$(BUILDDIR)', 'ngl-tools')),
+        _rd(op.join('$(BUILDDIR)', 'pkgconf')),  # XXX backport
+        _rd(op.join('$(BUILDDIR)', 'shaderc')),
+        _rd(op.join('$(BUILDDIR)', 'sxplayer')),
+        _rd(op.join('$(BUILDDIR)', 'tests')),
     ]
 
 
 def _coverage(cfg, output):
     # We don't use `meson coverage` here because of
     # https://github.com/mesonbuild/meson/issues/7895
-    return [_cmd_join('ninja', '-C', op.join('builddir', 'libnodegl'), f'coverage-{output}')]
+    return [_cmd_join('ninja', '-C', op.join('$(BUILDDIR)', 'libnodegl'), f'coverage-{output}')]
 
 
 @_block('coverage-html')
@@ -379,7 +480,7 @@ def _coverage_xml(cfg):
 
 @_block('tests', [_nodegl_tests, _tests_setup])
 def _tests(cfg):
-    return [_cmd_join('meson', 'test', '-C', op.join('builddir', 'tests'))]
+    return [_cmd_join('meson', 'test', '-C', op.join('$(BUILDDIR)', 'tests'))]
 
 
 @_block('shaderc-setup')
@@ -453,13 +554,15 @@ def _get_make_vars(cfg):
     # mechanism.
     #
     meson = 'MAKEFLAGS= meson' if _SYSTEM != 'Windows' else 'meson'
-
+    cmake = 'cmake'
+    builddir = os.getenv("BUILDDIR", 'builddir')
+    buildtype = 'debug' if cfg.args.coverage or cfg.args.buildtype == 'debug' else 'release'
     meson_setup = [
         'setup',
-        '--prefix', cfg.prefix,
+        '--prefix', '$(PREFIX)',
         '-Drpath=true',
-        '--pkg-config-path', cfg.pkg_config_path,
-        '--buildtype', 'debugoptimized' if debug else 'release',
+        '--pkg-config-path', '$(PKG_CONFIG_PATH)',
+        '--buildtype', buildtype,
     ]
     if cfg.args.coverage:
         meson_setup += ['-Db_coverage=true']
@@ -486,15 +589,31 @@ def _get_make_vars(cfg):
         meson_setup += ['-Db_vscrt=md']
 
     ret = dict(
-        PIP=_cmd_join(python, '-m', 'pip'),
-        MESON=meson,
+        BUILDDIR = builddir,
+        PREFIX = cfg.prefix, #.replace('\\','/'),
+        PIP = _cmd_join(python, '-m', 'pip'),
+        MESON = meson,
+        MESON_BACKEND = {'Windows': 'vs', 'Linux': 'ninja', 'Darwin': 'ninja', 'MinGW': 'ninja'}[_SYSTEM],
+        MESON_SETUP = '$(MESON) ' + _cmd_join(*meson_setup, '--backend=$(MESON_BACKEND)'),
+        MESON_SETUP_TESTS = '$(MESON) ' + _cmd_join(*meson_setup, '--backend=ninja'),
+        CMAKE = cmake,
+        CMAKE_GENERATOR = cfg.cmake_generator,
+        CMAKE_BUILD_TYPE = cfg.cmake_build_type,
     )
 
+    # FIXME: ...
     ret['MESON_SETUP'] = '$(MESON) ' + _cmd_join(*meson_setup, f'--backend={cfg.args.build_backend}')
     # Our tests/meson.build logic is not well supported with the VS backend so
     # we need to fallback on Ninja
     ret['MESON_SETUP_TESTS'] = '$(MESON) ' + _cmd_join(*meson_setup, '--backend=ninja')
     ret['CMAKE_SETUP'] = 'cmake ' + _cmd_join(*cmake_setup)
+
+    if _SYSTEM == 'Windows':
+        ret['CMAKE_SYSTEM_VERSION'] = cfg.cmake_system_version
+        ret['VCPKG_DIR'] = cfg.args.vcpkg_dir
+
+    if cfg.args.enable_ngfx_backend:
+        ret['NGFX_GRAPHICS_BACKEND'] = cfg.ngfx_graphics_backend
 
     return ret
 
@@ -579,13 +698,29 @@ class _Config:
         self.bin_path = op.join(self.prefix, self.bin_name)
         if _SYSTEM == 'Windows':
             vcpkg_pc_path = op.join(args.vcpkg_dir, 'installed', 'x64-windows', 'lib', 'pkgconfig')
-            self.pkg_config_path = os.pathsep.join((vcpkg_pc_path, op.join(self.prefix, 'Lib', 'pkgconfig')))
+            self.pkg_config_path = os.pathsep.join((vcpkg_pc_path, op.join(self.prefix, 'Lib', 'pkgconfig'))) #.replace('\\','/')
         else:
             self.pkg_config_path = op.join(self.prefix, 'lib', 'pkgconfig')
         self.externals = _fetch_externals(args)
 
+        if args.enable_ngfx_backend:
+            self.ngfx_graphics_backend = os.getenv('NGFX_GRAPHICS_BACKEND',
+                'NGFX_GRAPHICS_BACKEND_' + {'Windows': 'DIRECT3D12', 'Linux': 'VULKAN', 'Darwin': 'METAL'}[_SYSTEM])
+
+        self.cmake_generator = os.getenv("CMAKE_GENERATOR", {'Windows': 'Visual Studio 16 2019', 'Linux': 'Ninja', 'Darwin': 'Xcode', 'MinGW': 'Ninja'}[_SYSTEM])
+
+        if _SYSTEM == 'Windows':
+            # Set Windows SDK Version
+            self.cmake_system_version = os.getenv('CMAKE_SYSTEM_VERSION', '10.0.18362.0')
+
+        if args.buildtype == 'debug':
+            self.cmake_build_type = 'Debug'
+        else:
+            self.cmake_build_type = 'Release'
+
         if _SYSTEM == 'Windows':
             _sxplayer_setup.prerequisites.append(_pkgconf_install)
+            _ngfx_install.prerequisites.insert(0, _pkgconf_install)
             if 'gpu_capture' in args.debug_opts:
                 _nodegl_setup.prerequisites.append(_renderdoc_install)
 
@@ -593,20 +728,25 @@ class _Config:
             for f in glob.glob(op.join(vcpkg_bin, '*.dll')):
                 logging.info('copy %s to venv/Scripts', f)
                 shutil.copy2(f, op.join('venv', 'Scripts'))
+
             shaderc_dll = op.join(args.vcpkg_dir, 'installed', 'x64-windows', 'tools', 'shaderc_shared.dll')
             if op.exists(shaderc_dll):
                 shutil.copy2(shaderc_dll, op.join('venv', 'Scripts'))
+
+        if args.enable_ngfx_backend:
+            _nodegl_setup.prerequisites.append(_ngfx_install)
 
     def get_env(self):
         sep = ':' if _SYSTEM == 'MinGW' else os.pathsep
         env = {}
         env['PATH'] = sep.join((self.bin_path, '$(PATH)'))
+        env['EXTERNAL_DIR'] = op.join(_ROOTDIR, 'external')
         env['PKG_CONFIG_PATH'] = self.pkg_config_path
         if _SYSTEM == 'Windows':
+            env['PKG_CONFIG'] = op.join(self.args.venv_path, 'Scripts', 'pkgconf.exe')
             env['PKG_CONFIG_ALLOW_SYSTEM_LIBS'] = '1'
             env['PKG_CONFIG_ALLOW_SYSTEM_CFLAGS'] = '1'
         return env
-
 
 def _run():
     default_build_backend = 'ninja' if _SYSTEM != 'Windows' else 'vs'
@@ -625,9 +765,13 @@ def _run():
                         help='Debug options')
     parser.add_argument('--build-backend', choices=('ninja', 'vs'), default=default_build_backend,
                         help='Build backend to use')
+    parser.add_argument('-v','--verbose', action='store_true',
+                        help='Enable verbose output')
     if _SYSTEM == 'Windows':
         parser.add_argument('--vcpkg-dir', default=r'C:\vcpkg',
                             help='Vcpkg directory')
+    parser.add_argument('--enable-ngfx-backend', action='store_true',
+                        help='Enable NGFX Backend')
 
     args = parser.parse_args()
 
@@ -646,10 +790,10 @@ def _run():
         blocks += [_shaderc_install, _moltenvk_install]
     if args.coverage:
         blocks += [_coverage_html, _coverage_xml]
+
     makefile = _get_makefile(cfg, blocks)
     with open(dst_makefile, 'w') as f:
         f.write(makefile)
-
 
 if __name__ == '__main__':
     _run()
